Renamed dossier-management-list
