+++ projects/valtimo/case-management/src/lib/components/case-management-list/case-management-list.component.html
@@ -47,7 +47,7 @@
 ></valtimo-case-management-create>
 
 <ng-template #importCaseDefinitionButton>
-  <div class="valtimo-dossier-management-list__actions">
+  <div class="valtimo-case-management-list__actions">
     <button cdsButton="ghost" disabled="true" (click)="showUploadModal()">
       {{ 'Upload' | translate }}
