+++ projects/valtimo/case-management/src/lib/components/case-management-list/case-management-list.component.scss
@@ -13,6 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-.valtimo-dossier-management-list__actions {
+.valtimo-case-management-list__actions {
   display: flex;
 }
